refactor: rename `/examples/` to `/_examples/`
